Add HRU 250/300 Fan + Rename base files
