+++ custom_components/ithodaalderop/definitions/co2_remote.py
@@ -3,7 +3,7 @@
 from homeassistant.components.sensor import SensorDeviceClass, SensorStateClass
 from homeassistant.const import CONCENTRATION_PARTS_PER_MILLION
 
-from .base import IthoSensorEntityDescription
+from .base_definitions import IthoSensorEntityDescription
 
 REMOTE_SENSOR_TEMPLATE = IthoSensorEntityDescription(
     key="remote",
