+++ custom_components/ithodaalderop/definitions/demandflow.py
@@ -9,7 +9,10 @@ from homeassistant.const import (
     UnitOfVolumeFlowRate,
 )
 
-from .base import IthoBinarySensorEntityDescription, IthoSensorEntityDescription
+from .base_definitions import (
+    IthoBinarySensorEntityDescription,
+    IthoSensorEntityDescription,
+)
 
 DEMAND_FLOW_BINARY_SENSORS: tuple[IthoBinarySensorEntityDescription, ...] = (
     IthoBinarySensorEntityDescription(
