+++ custom_components/ithodaalderop/definitions/hrueco.py
@@ -10,7 +10,10 @@ from homeassistant.const import (
     UnitOfTime,
 )
 
-from .base import IthoBinarySensorEntityDescription, IthoSensorEntityDescription
+from .base_definitions import (
+    IthoBinarySensorEntityDescription,
+    IthoSensorEntityDescription,
+)
 
 HRU_ECO_BINARY_SENSORS: tuple[IthoBinarySensorEntityDescription, ...] = (
     IthoBinarySensorEntityDescription(
