+++ custom_components/ithodaalderop/fans/_itho_HADiscovery.temp.cpp
