+++ custom_components/ithodaalderop/utils.py
@@ -10,6 +10,7 @@ from .const import (
     CONF_CUSTOM_DEVICE_NAME,
     CONF_CUSTOM_ENTITY_PREFIX,
     CONF_NONCVE_MODEL,
+    MQTT_COMMAND_TOPIC,
     MQTT_DEFAULT_BASETOPIC,
     MQTT_STATETOPIC,
     NONCVE_DEVICES,
@@ -27,6 +28,11 @@ def get_mqtt_base_topic(config: dict[str, Any]) -> str:
         return config[CONF_CUSTOM_BASETOPIC]
 
     return get_default_mqtt_base_topic(config)
+
+
+def get_mqtt_command_topic(config: dict[str, Any]) -> str:
+    """Get the MQTT command topic."""
+    return f"{get_mqtt_base_topic(config)}/{MQTT_COMMAND_TOPIC}"
 
 
 def get_mqtt_state_topic(config: dict[str, Any]) -> str:
